Move Inversed to a separate file
